Try out travis with running indexing and the quality tests
A new `test:quality:exit0` script exists to make the overall exit code of the script 0.  That should prevent the red "x" from Travis on the quality tests.  We want to get as many passing as possible, but it's a moving target.

The ideal would be to automate tracking of the failure count, but let's review this count manually for now.